requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.1
motor==3.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
spotipy==2.23.0
python-dotenv==1.0.0
pydantic[email]==2.5.2
pydantic-settings==2.1.0
cryptography==41.0.7
openai==1.3.5
python-jose==3.3.0
email-validator==2.1.0.post1
bcrypt==4.0.1 